Refactor container utils into separate class.
