+++ src/main/java/bio/terra/cli/apps/AppsRunner.java
@@ -4,28 +4,8 @@ import bio.terra.cli.command.exception.SystemException;
 import bio.terra.cli.context.GlobalContext;
 import bio.terra.cli.context.TerraUser;
 import bio.terra.cli.context.WorkspaceContext;
-import bio.terra.cli.context.utils.Printer;
-import com.github.dockerjava.api.DockerClient;
-import com.github.dockerjava.api.async.ResultCallback;
-import com.github.dockerjava.api.command.CreateContainerCmd;
-import com.github.dockerjava.api.command.CreateContainerResponse;
-import com.github.dockerjava.api.command.WaitContainerResultCallback;
-import com.github.dockerjava.api.exception.NotFoundException;
-import com.github.dockerjava.api.model.Bind;
-import com.github.dockerjava.api.model.Frame;
-import com.github.dockerjava.api.model.HostConfig;
-import com.github.dockerjava.api.model.Volume;
-import com.github.dockerjava.core.DefaultDockerClientConfig;
-import com.github.dockerjava.core.DockerClientConfig;
-import com.github.dockerjava.core.DockerClientImpl;
-import com.github.dockerjava.httpclient5.ApacheDockerHttpClient;
-import com.github.dockerjava.transport.DockerHttpClient;
-import java.io.File;
-import java.io.PrintWriter;
-import java.nio.charset.Charset;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -36,12 +16,12 @@ import org.slf4j.LoggerFactory;
  * This class runs client-side tools and manipulates the tools-related properties of the global
  * context object.
  */
-public class DockerAppsRunner {
-  private static final Logger logger = LoggerFactory.getLogger(DockerAppsRunner.class);
+public class AppsRunner {
+  private static final Logger logger = LoggerFactory.getLogger(AppsRunner.class);
 
   private final GlobalContext globalContext;
   private final WorkspaceContext workspaceContext;
-  private DockerClient dockerClient;
+  private final DockerUtils dockerUtils = new DockerUtils();
 
   // default $HOME directory on the container (this is where we expect to look for the global
   // context)
@@ -49,10 +29,9 @@ public class DockerAppsRunner {
   // mount point for the workspace directory
   private static final String CONTAINER_WORKSPACE_DIR = "/usr/local/etc";
 
-  public DockerAppsRunner(GlobalContext globalContext, WorkspaceContext workspaceContext) {
+  public AppsRunner(GlobalContext globalContext, WorkspaceContext workspaceContext) {
     this.globalContext = globalContext;
     this.workspaceContext = workspaceContext;
-    this.dockerClient = buildDockerClient();
   }
 
   // ====================================================
@@ -61,7 +40,7 @@ public class DockerAppsRunner {
   /** Returns the default image id. */
   public static String defaultImageId() {
     // read from the JAR Manifest file
-    return DockerAppsRunner.class.getPackage().getImplementationVersion();
+    return AppsRunner.class.getPackage().getImplementationVersion();
   }
 
   /**
@@ -73,11 +52,9 @@ public class DockerAppsRunner {
    * @param imageId id or tag of the image
    */
   public void updateImageId(String imageId) {
-    // check if image exists
-    try {
-      dockerClient.inspectImageCmd(imageId).exec();
-    } catch (NotFoundException nfEx) {
-      logger.warn("Image not found: {}", imageId, nfEx);
+    boolean imageExists = dockerUtils.checkImageExists(imageId);
+    if (!imageExists) {
+      logger.warn("Image not found: {}", imageId);
     }
 
     globalContext.updateDockerImageId(imageId);
@@ -153,14 +130,14 @@ public class DockerAppsRunner {
     String containerId = startDockerContainerWithTerraInit(command, envVars, bindMounts);
 
     // read the container logs, which contains the command output, and write them to stdout
-    outputLogsForDockerContainer(containerId);
+    dockerUtils.outputLogsForDockerContainer(containerId);
 
     // block until the container exits
-    Integer statusCode = waitForDockerContainerToExit(containerId);
+    Integer statusCode = dockerUtils.waitForDockerContainerToExit(containerId);
     logger.debug("docker run status code: {}", statusCode);
 
     // delete the container
-    deleteDockerContainer(containerId);
+    dockerUtils.deleteDockerContainer(containerId);
   }
 
   /**
@@ -225,153 +202,12 @@ public class DockerAppsRunner {
     }
     envVars.putAll(terraInitEnvVars);
 
-    return startDockerContainer(
-        fullCommand, getWorkingDirOnContainer().toString(), envVars, bindMounts);
-  }
-
-  // ====================================================
-  // Docker containers
-
-  /** Build the Docker client object with standard options. */
-  private static DockerClient buildDockerClient() {
-    DockerClientConfig config = DefaultDockerClientConfig.createDefaultConfigBuilder().build();
-    DockerHttpClient httpClient =
-        new ApacheDockerHttpClient.Builder()
-            .dockerHost(config.getDockerHost())
-            .sslConfig(config.getSSLConfig())
-            .build();
-    return DockerClientImpl.getInstance(config, httpClient);
-  }
-
-  /**
-   * Start a Docker container and run the given command. Return the container id.
-   *
-   * @param command the full string command to execute in a bash shell (bash -c ..cmd..)
-   * @param workingDir the directory where the commmand will be executed
-   * @param envVars a mapping of environment variable names to values
-   * @param bindMounts a mapping of container mount point to the local directory being mounted
-   * @throws SystemException if the local directory does not exist or is not a directory
-   */
-  private String startDockerContainer(
-      String command, String workingDir, Map<String, String> envVars, Map<Path, Path> bindMounts) {
-    // flatten the environment variables from a map, into a list of key=val strings
-    List<String> envVarsStr = new ArrayList<>();
-    for (Map.Entry<String, String> envVar : envVars.entrySet()) {
-      envVarsStr.add(envVar.getKey() + "=" + envVar.getValue());
-    }
-
-    // create Bind objects for each specified mount
-    List<Bind> bindMountsObj = new ArrayList<>();
-    for (Map.Entry<Path, Path> bindMount : bindMounts.entrySet()) {
-      File localDirectory = bindMount.getValue().toFile();
-      if (!localDirectory.exists() || !localDirectory.isDirectory()) {
-        throw new SystemException(
-            "Bind mount does not specify a local directory: " + localDirectory.getAbsolutePath());
-      }
-      bindMountsObj.add(
-          new Bind(localDirectory.getAbsolutePath(), new Volume(bindMount.getKey().toString())));
-    }
-
-    // create the container and start it
-    CreateContainerCmd createContainerCmd =
-        dockerClient
-            .createContainerCmd(globalContext.dockerImageId)
-            .withCmd("bash", "-c", command)
-            .withEnv(envVarsStr)
-            .withHostConfig(HostConfig.newHostConfig().withBinds(bindMountsObj))
-            .withAttachStdout(true)
-            .withAttachStderr(true);
-    if (workingDir != null) {
-      createContainerCmd.withWorkingDir(workingDir);
-    }
-    CreateContainerResponse container = createContainerCmd.exec();
-    dockerClient.startContainerCmd(container.getId()).exec();
-    return container.getId();
-  }
-
-  /**
-   * Block until the Docker container exits, then return its status code.
-   *
-   * @param containerId id of the container
-   */
-  private Integer waitForDockerContainerToExit(String containerId) {
-    WaitContainerResultCallback waitContainerResultCallback = new WaitContainerResultCallback();
-    WaitContainerResultCallback exec =
-        dockerClient
-            .waitContainerCmd(containerId)
-            .<WaitContainerResultCallback>exec(waitContainerResultCallback);
-    return exec.awaitStatusCode();
-  }
-
-  /**
-   * Delete the Docker container.
-   *
-   * @param containerId id of the container
-   */
-  private void deleteDockerContainer(String containerId) {
-    dockerClient.removeContainerCmd(containerId).exec();
-  }
-
-  /**
-   * Read the Docker container logs and write them to standard out.
-   *
-   * @param containerId id of the container
-   */
-  private void outputLogsForDockerContainer(String containerId) {
-    dockerClient
-        .logContainerCmd(containerId)
-        .withStdOut(true)
-        .withStdErr(true)
-        .withFollowStream(true)
-        .withTailAll()
-        .exec(new DockerAppsRunner.LogContainerTestCallback());
-  }
-
-  /** Helper class for reading Docker container logs into a string. */
-  private static class LogContainerTestCallback extends ResultCallback.Adapter<Frame> {
-
-    protected final StringBuffer log = new StringBuffer();
-    List<Frame> framesList = new ArrayList<>();
-
-    // these two boolean flags are useful for debugging
-    // buildSingleStringOutput = concatenate the output into a single String (be careful of very
-    // long outputs)
-    boolean buildSingleStringOutput;
-    // buildFramesList = keep a list of all the output lines (frames) as they come back
-    boolean buildFramesList;
-
-    public LogContainerTestCallback() {
-      this(false, false);
-    }
-
-    public LogContainerTestCallback(boolean buildSingleStringOutput, boolean buildFramesList) {
-      this.buildSingleStringOutput = buildSingleStringOutput;
-      this.buildFramesList = buildFramesList;
-    }
-
-    @Override
-    public void onNext(Frame frame) {
-      String logStr = new String(frame.getPayload(), Charset.forName("UTF-8"));
-      PrintWriter err = Printer.getErr();
-      err.print(logStr);
-      err.flush();
-
-      if (buildSingleStringOutput) {
-        log.append(logStr);
-      }
-      if (buildFramesList) {
-        framesList.add(frame);
-      }
-    }
-
-    @Override
-    public String toString() {
-      return log.toString();
-    }
-
-    public List<Frame> getFramesList() {
-      return framesList;
-    }
+    return dockerUtils.startDockerContainer(
+        globalContext.dockerImageId,
+        fullCommand,
+        getWorkingDirOnContainer().toString(),
+        envVars,
+        bindMounts);
   }
 
   // ====================================================
